CCS-68 refactor: 외부 API 요청 클래스 네이밍, 패키지 이름 변경
